forge fmt

+++ src/ProofOfBurnVerifier.sol
@@ -22,17 +22,17 @@ pragma solidity >=0.7.0 <0.9.0;
 
 contract ProofOfBurnVerifier {
     // Scalar field size
-    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant alphax  = 20491192805390485299153009773594534940189261866228447918068658471970481763042;
-    uint256 constant alphay  = 9383485363053290200918347156157836566562967994039712273449902621266178545958;
-    uint256 constant betax1  = 4252822878758300859123897981450591353533073413197771768651442665752259397132;
-    uint256 constant betax2  = 6375614351688725206403948262868962793625744043794305715222011528459656738731;
-    uint256 constant betay1  = 21847035105528745403288232691147584728191162732299865338377159692350059136679;
-    uint256 constant betay2  = 10505242626370262277552901082094356697409835680220590971873171140371331206856;
+    uint256 constant alphax = 20491192805390485299153009773594534940189261866228447918068658471970481763042;
+    uint256 constant alphay = 9383485363053290200918347156157836566562967994039712273449902621266178545958;
+    uint256 constant betax1 = 4252822878758300859123897981450591353533073413197771768651442665752259397132;
+    uint256 constant betax2 = 6375614351688725206403948262868962793625744043794305715222011528459656738731;
+    uint256 constant betay1 = 21847035105528745403288232691147584728191162732299865338377159692350059136679;
+    uint256 constant betay2 = 10505242626370262277552901082094356697409835680220590971873171140371331206856;
     uint256 constant gammax1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
     uint256 constant gammax2 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
     uint256 constant gammay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
@@ -42,21 +42,24 @@ contract ProofOfBurnVerifier {
     uint256 constant deltay1 = 20184189169096204033875621574152617736805009299164766169960708720711561347485;
     uint256 constant deltay2 = 12227578378264111381899415289511317458920259146968670324680247968202123849226;
 
-    
     uint256 constant IC0x = 17981016415722245772442382858927564259834542968495105279336873778045267855424;
     uint256 constant IC0y = 6727694969926707052939778545606762903599824719387877996827866518158636831549;
-    
+
     uint256 constant IC1x = 19737023379765774210194385307893707984716381403776209139995368476453787818080;
     uint256 constant IC1y = 18126007243504095946874613321957771918526426501936487307392985066864586050295;
-    
- 
+
     // Memory data
     uint16 constant pVk = 0;
     uint16 constant pPairing = 128;
 
     uint16 constant pLastMem = 896;
 
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[1] calldata _pubSignals) public view returns (bool) {
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[1] calldata _pubSignals
+    ) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, r)) {
@@ -64,7 +67,7 @@ contract ProofOfBurnVerifier {
                     return(0, 0x20)
                 }
             }
-            
+
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -99,9 +102,8 @@ contract ProofOfBurnVerifier {
                 mstore(add(_pVk, 32), IC0y)
 
                 // Compute the linear combination vk_x
-                
+
                 g1_mulAccC(_pVk, IC1x, IC1y, calldataload(add(pubSignals, 0)))
-                
 
                 // -A
                 mstore(_pPairing, calldataload(pA))
@@ -127,7 +129,6 @@ contract ProofOfBurnVerifier {
                 mstore(add(_pPairing, 384), mload(add(pMem, pVk)))
                 mstore(add(_pPairing, 416), mload(add(pMem, add(pVk, 32))))
 
-
                 // gamma2
                 mstore(add(_pPairing, 448), gammax1)
                 mstore(add(_pPairing, 480), gammax2)
@@ -144,7 +145,6 @@ contract ProofOfBurnVerifier {
                 mstore(add(_pPairing, 704), deltay1)
                 mstore(add(_pPairing, 736), deltay2)
 
-
                 let success := staticcall(sub(gas(), 2000), 8, _pPairing, 768, _pPairing, 0x20)
 
                 isOk := and(success, mload(_pPairing))
@@ -154,15 +154,14 @@ contract ProofOfBurnVerifier {
             mstore(0x40, add(pMem, pLastMem))
 
             // Validate that all evaluations ∈ F
-            
+
             checkField(calldataload(add(_pubSignals, 0)))
-            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-             return(0, 0x20)
-         }
-     }
- }
+            return(0, 0x20)
+        }
+    }
+}
